+++ CHANGELOG.md
@@ -4,6 +4,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased][unreleased]
 ### Added
+ - Nothing yet
+
+
+## [v0.8.0] - 2016-11-11
+### Added
  - `Release` section to the README
  - Fastboot support
  - Nested addon support
@@ -168,7 +173,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
  - Initial transformicons
 
 
-[unreleased]:  https://github.com/alexdiliberto/ember-cli-transformicons/compare/v0.7.0...HEAD
+[unreleased]:  https://github.com/alexdiliberto/ember-cli-transformicons/compare/v0.8.0...HEAD
+[v0.8.0]:      https://github.com/alexdiliberto/ember-cli-transformicons/compare/v0.7.0...v0.8.0
 [v0.7.0]:      https://github.com/alexdiliberto/ember-cli-transformicons/compare/v0.6.0...v0.7.0
 [v0.6.0]:      https://github.com/alexdiliberto/ember-cli-transformicons/compare/v0.5.0...v0.6.0
 [v0.5.0]:      https://github.com/alexdiliberto/ember-cli-transformicons/compare/v0.4.0...v0.5.0